requirements.txt
#abjad==3.13
celery==5.2.7
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers
#django-extensions==3.2.1
google-cloud-translate==2.0.1
gunicorn==20.1.0
#mingus==0.6.1
music21==8.1.0
#nltk==3.7
parce==0.32.0
python-dotenv==0.21.0
PyYAML==6.0
#quickly==0.7.0
redis==4.3.4
spacy==3.4.1
uvicorn==0.18.3
